imoproved authorized user delete

+++ src/controllers/questions.js
@@ -1,10 +1,10 @@
 import QuestionsModel from '../models/questions.js'
-
+import AnswersModel from '../models/answers.js';
+import UsersModel from '../models/users.js';
 const ADD_QUESTION = async (req, res) => {
     try {
         const question = new QuestionsModel({
             question_text: req.body.question_text,
-            // answer_id: req.body.answer_id,
             user_id: req.body.userId,
         });
         question.id = question._id
@@ -15,6 +15,24 @@ const ADD_QUESTION = async (req, res) => {
         return res.status(500).json({ status: "Error occurred" });
     }
 };
+const GET_QUESTIONS = async (req, res) => {
+    try {
+        const questions = await QuestionsModel.find()
+        return res.status(200).json({ questions, status: "Questions" })
+    } catch (err) {
+        console.log(err)
+        return res.status(500).json({ status: "Error ocurred", })
+    }
+}
+const GET_QUESTION_ID = async (req, res) => {
+    try {
+        const question = await QuestionsModel.findById(req.params.id)
+        return res.status(200).json({ question, status: "Questions" })
+    } catch (err) {
+        console.log(err)
+        return res.status(500).json({ status: "Error ocurred", })
+    }
+}
 const GET_QUESTION_ANSWER = async (req, res) => {
     try {
 
@@ -32,7 +50,7 @@ const GET_QUESTION_ANSWER = async (req, res) => {
                 $match: { id: (req.params.id) } // Match the question by id
             }
         ]);
-        console.log(questionAnswer)
+
         return res.status(200).json({ questionAnswer, status: "Questions answer" })
     } catch (err) {
         console.log(err)
@@ -113,17 +131,26 @@ const GET_QUESTION_WITH_NO_ANSWERS = async (req, res) => {
     }
 };
 
+
+
 const DELETE_QUESTION = async (req, res) => {
     try {
-        const questionDelete = await QuestionsModel.findByIdAndDelete(req.params.id)
-        if (!questionDelete) {
-            return res.status(404).json({ status: "Question not exist" });
-        }
-        return res.status(200).json({ response: questionDelete, status: "Question was deleted" });
-    } catch (err) {
-        console.log(err)
-        return res.status(500).json({ status: "Error ocurred", })
-    }
-}
+        const question = await QuestionsModel.findById(req.params.id);
+        const user = await UsersModel.findById(req.body.userId);
+        const user_id = user.id;
 
-export { ADD_QUESTION, GET_QUESTIONS_ANSWERS, DELETE_QUESTION, GET_QUESTION_ANSWER, GET_QUESTION_WITH_NO_ANSWERS, GET_QUESTION_WITH_ANSWERS }
+        if (question.user_id == user_id) {
+            const deleted = await QuestionsModel.findByIdAndDelete(req.params.id)
+            return res.status(200).json({ deleted, message: "Comment deleted successfully" });
+        } else {
+            return res.status(401).json({ message: "Unauthorized to delete this comment" });
+        }
+
+    } catch (error) {
+        console.error(error);
+        return res.status(500).json({ message: "Internal server error" });
+    }
+};
+
+
+export { ADD_QUESTION, GET_QUESTIONS_ANSWERS, DELETE_QUESTION, GET_QUESTION_ANSWER, GET_QUESTION_WITH_NO_ANSWERS, GET_QUESTION_WITH_ANSWERS, GET_QUESTIONS, GET_QUESTION_ID }
